move starboard to events folder
